Move diagnostics list
