+++ Composer/packages/client/__tests__/pages/notifications/diagnosticList.test.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { Range, Position } from '@bfc/shared';
 
 import { renderWithRecoil } from '../../testUtils';
-import { DiagnosticList } from '../../../src/pages/diagnostics/DiagnosticList';
+import { DiagnosticList } from '../../../src/pages/design/DebugPanel/TabExtensions/DiagnosticsTab/DiagnosticList';
 import {
   botDiagnosticsState,
   botProjectIdsState,
